+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v4.23.0
+
+### Added
+
+- Add methods `trimPrefix`, `trimSuffix` into `Str`
+
 ## v4.22.0
 
 ### Added